modify error in ewpsp

+++ ewpsp/src/main/java/kr/co/ewp/ewpsp/service/KepcoMngSetServiceImpl.java
@@ -31,10 +31,10 @@ public class KepcoMngSetServiceImpl implements KepcoMngSetService {
 	
 	@Transactional
 	public int updateSiteSet(HashMap param) throws Exception {
-		if(param.get("essBattery") != null) param.put("essBattery", ( Integer.parseInt((String)param.get("essBattery")) )*1000); // kWh -> Wh
-		if(param.get("essPcs") != null) param.put("essPcs", ( Integer.parseInt((String)param.get("essPcs")) )*1000); // kW -> W
-		if(param.get("contractPower") != null) param.put("contractPower", ( Integer.parseInt((String)param.get("contractPower")) )*1000); // kWh -> Wh
-		if(param.get("chargePower") != null) param.put("chargePower", ( Integer.parseInt((String)param.get("chargePower")) )*1000); // kWh -> Wh
+		if(param.get("essBattery") != null) param.put("essBattery", ( Float.parseFloat((String)param.get("essBattery")) )*1000); // kWh -> Wh
+		if(param.get("essPcs") != null) param.put("essPcs", ( Float.parseFloat((String)param.get("essPcs")) )*1000); // kW -> W
+		if(param.get("contractPower") != null) param.put("contractPower", ( Float.parseFloat((String)param.get("contractPower")) )*1000); // kWh -> Wh
+		if(param.get("chargePower") != null) param.put("chargePower", ( Float.parseFloat((String)param.get("chargePower")) )*1000); // kWh -> Wh
 		return kepcoMngSetDao.updateSiteSet(param);
 	}
